refactor(cmd): split actor cli module
